requirements.txt
pandas
qrcode[pil]
Pillow
